requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.0
python-docx>=1.1.0
scikit-learn>=1.4.0
google-generativeai>=0.5.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
